refactor: add stdout writer factory and json implementation
